refactor(P2.5 sub-slice D): move LogicalPlan-level CTE rewriters into render_plan/cte_graph_joins_rewrite.rs
Phase-2 §5.1 module move, fourth/final function sub-slice of the cte_rewrite
group. Extracted the contiguous LogicalPlan-level CTE-ref rewrite block verbatim
from plan_builder_utils.rs to a NEW render_plan/cte_graph_joins_rewrite.rs:
rewrite_logical_expr_cte_refs, the 3 alias walkers deferred from P2.4
(find_fresh_table_scan_aliases_in_plan / collect_fresh_scan_aliases /
with_exported_aliases_in_branch), and the ~400-line update_graph_joins_cte_refs
(old lines 3547-4120, ~574 lines).

Kept as a separate module from cte_rewrite.rs because it operates on
query_planner::logical_expr / logical_plan types, not render types (the
LogicalPlan companion). Only update_graph_joins_cte_refs has in-module callers
(2 sites, already pub(crate)), so it alone is re-exported; the other 4 are called
solely within the new module and stay private.

Fully byte-identical — zero visibility changes needed (unlike sub-slices A/C).
Ruled out a false-positive external caller (Self::collect_fresh_scan_aliases in
analyzer/graph_join/inference.rs is an unrelated same-named method). This
completes the P2.5 function moves; only the D2 dedup remains.

Gates: 1612 lib + 529 integration (incl. corpus_sweep + sql_golden) + 1 ratchet
+ 7 unit, 0 failures; byte-identical (moved block diff-clean vs main, no golden
churn, ratchet net-zero); fmt/clippy clean; warning-free --tests build.
